Refactor run-script-if to allow multiple conditions
