+++ packages/image-builder/package.json
@@ -22,7 +22,10 @@
     "src"
   ],
   "scripts": {
-    "install": "rimraf dist && tsc"
+    "install": "pnpm write-readme && rimraf dist && tsc",
+    "write-readme": "run-script-os",
+    "write-readme:darwin:linux": "echo '```' > README.md && ./bin.js --help >> README.md && echo '```' >> README.md",
+    "write-readme:win32": "echo 'Nothing to do.'"
   },
   "dependencies": {
     "@kie-tools-scripts/build-env": "workspace:*",
@@ -31,6 +34,7 @@
   "devDependencies": {
     "@types/yargs": "^17.0.1",
     "rimraf": "^3.0.2",
+    "run-script-os": "^1.1.6",
     "typescript": "^4.6.2"
   }
 }
