+++ packages/image-builder/src/bin.ts
@@ -1,3 +1,21 @@
+#!/usr/bin/env node
+
+/*
+ * Copyright 2023 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 import * as path from "path";
 import { execSync } from "child_process";
 import { findEnv, flattenObj } from "@kie-tools-scripts/build-env";
@@ -5,8 +23,8 @@ import * as yargs from "yargs";
 import { hideBin } from "yargs/helpers";
 
 async function main() {
-  const { self } = await findEnv(path.resolve("."), path.resolve("."));
-  const flattenedEnv = flattenObj(self.env);
+  const { env } = await findEnv(path.resolve("."), path.resolve("."));
+  const flattenedEnv = flattenObj(env);
 
   const parseBuildEnvPath = (path: string) => {
     if (!flattenedEnv[path]) {
