+++ packages/kie-sandbox-distribution/package.json
@@ -14,14 +14,14 @@
   },
   "scripts": {
     "build:dev": "pnpm docker:build",
-    "build:prod": "run-script-if --bool \"$(build-env containerImages.build)\" --then \"pnpm docker:build\" \"pnpm test\"",
+    "build:prod": "run-script-if --build-env \"containerImages.build\" --then \"pnpm docker:build\" \"pnpm test\"",
     "docker:build": "docker compose build",
     "docker:create-env-file": "rimraf .env && pnpm build-env --print-env-file:self > .env",
     "docker:down": "docker compose down",
     "docker:start": "docker compose up -d --wait",
     "docker:stop": "docker compose stop",
     "install": "pnpm docker:create-env-file",
-    "test": "run-script-if --ignore-errors \"$(build-env tests.ignoreFailures)\" --bool \"$(node scripts/shouldRunTest.js)\" --then \"pnpm docker:start\" \"node scripts/test.js\" \"pnpm docker:stop\""
+    "test": "run-script-if --ignore-errors \"$(build-env tests.ignoreFailures)\" --build-env \"containerImages.build\" --build-env \"tests.run\" --then  \"pnpm docker:start\" \"node scripts/test.js\" \"pnpm docker:stop\""
   },
   "dependencies": {
     "@kie-tools/git-cors-proxy-image": "workspace:*",
