--- packages/kie-sandbox-distribution/scripts/shouldRunTest.js
@@ -1,8 +1,0 @@
-const kieSandboxDistributionEnv = require("../env");
-
-const env = kieSandboxDistributionEnv.env;
-
-const buildContainerImages = env.containerImages.build;
-const runTests = env.tests.run;
-
-console.log(buildContainerImages && runTests ? "true" : "false");
