+++ scripts/run-script-if/bin.js
