+++ scripts/run-script-if/package.json
@@ -13,21 +13,27 @@
     "url": "https://github.com/kiegroup/kie-tools/issues"
   },
   "bin": {
-    "run-script-if": "index.js"
+    "run-script-if": "bin.js"
   },
   "files": [
-    "index.js"
+    "bin.js",
+    "dist",
+    "src"
   ],
   "scripts": {
-    "install": "pnpm write-readme",
+    "install": "pnpm write-readme && rimraf dist && tsc",
     "write-readme": "run-script-os",
-    "write-readme:darwin:linux": "echo '```' > README.md && ./index.js --help >> README.md && echo '```' >> README.md",
+    "write-readme:darwin:linux": "echo '```' > README.md && ./bin.js --help >> README.md && echo '```' >> README.md",
     "write-readme:win32": "echo 'Nothing to do.'"
   },
   "dependencies": {
+    "@kie-tools-scripts/build-env": "workspace:*",
     "yargs": "^17.0.1"
   },
   "devDependencies": {
-    "run-script-os": "^1.1.6"
+    "@types/yargs": "^17.0.1",
+    "rimraf": "^3.0.2",
+    "run-script-os": "^1.1.6",
+    "typescript": "^4.6.2"
   }
 }
